requirements.txt
Django==2.2.1
psycopg2-binary==2.8.2
djangorestframework==3.9.3
shortuuid==0.5.0
PyJWT==1.7.1
